+++ src/Manager/nrfutil/terminal.ts
@@ -4,41 +4,40 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import { exec } from 'child_process';
-
 import { persistedInstallDir as installDir } from '../../persistentStore';
 import sdkPath from '../sdkPath';
+import { nrfutilExec } from './nrfutilChildProcess';
 import nrfutilToolchainManager from './nrfutilToolchainManager';
 
 export const launchWinBash = (version: string) => {
-    exec(
+    nrfutilExec(
         `"${nrfutilToolchainManager()}" launch --chdir "${sdkPath(
             version
         )}" --ncs-version "${version}" --install-dir "${installDir()}" cmd.exe /k start bash.exe`,
         {
-            env: { ...process.env, ZEPHYR_BASE: sdkPath(version, 'zephyr') },
+            ZEPHYR_BASE: sdkPath(version, 'zephyr'),
         }
     );
 };
 
 export const launchTerminal = (version: string) => {
-    exec(
+    nrfutilExec(
         `"${nrfutilToolchainManager()}" launch --chdir "${sdkPath(
             version
         )}" --ncs-version "${version}" --install-dir "${installDir()}" --terminal`,
         {
-            env: { ...process.env, ZEPHYR_BASE: sdkPath(version, 'zephyr') },
+            ZEPHYR_BASE: sdkPath(version, 'zephyr'),
         }
     );
 };
 
 export const launchGnomeTerminal = (version: string) => {
-    exec(
+    nrfutilExec(
         `gnome-terminal -- "${nrfutilToolchainManager()}" launch --chdir "${sdkPath(
             version
         )}" --ncs-version "${version}" --install-dir "${installDir()}" --shell`,
         {
-            env: { ...process.env, ZEPHYR_BASE: sdkPath(version, 'zephyr') },
+            ZEPHYR_BASE: sdkPath(version, 'zephyr'),
         }
     );
 };